Make custom notifications  custom forced inform and boot inform parameters in 3 separated files

+++ common.c
@@ -644,7 +644,7 @@ void icwmp_restart_services()
  */
 bool icwmp_validate_string_length(char *arg, int max_length)
 {
-	if (arg != NULL && strlen(arg) > max_length)
+	if (arg != NULL && strlen(arg) > (size_t)max_length)
 		return false;
 	return true;
 }
